fit and transform on list of images
